### 2024-08-22 #652 @heyairf
pre-submission inquiry: fireexposuR
Submitting Author Name: Air Forbes
Submitting Author Github Handle: @heyairf

Repository:  https://github.com/heyairf/fireexposuR
Submission type: Pre-submission

---

```
Package: fireexposuR
Title: Compute and Visualize Wildfire Exposure
Version: 0.2.0
Authors@R: c(person("Air", "Forbes", email = "amforbes@ualberta.ca",
    role = c("aut", "cre"), comment = c(ORCID = "0000-0002-9842-7648")),
    person("Jennifer", "Beverly", role = "aut"))
Description: This package computes wildfire exposure using methods from
    Beverly et al. (2010), Beverly et al. (2021), and Beverly and Forbes
    (2023).  It provides functions to standardize the mapping and
    visualization of wildfire exposure. This package requires
    pre-processing of data which is documented in Forbes and Beverly
    (Manuscript in preparation).
    geosphere,
    ggspatial,
    maptiles,
    MultiscaleDTM,
    terra,
    tidyselect,
    tidyterra
```

This package was developed to share methodologies from existing research by modifying/expanding existing functions. The functions automate and standardize assessments to increase access and quality assurance in the calculation for interested users.

n/a

Users interested in wildfire risk assessments including but not limited to researchers, consultants, planners, land use decision makers.

There are no other R packages that fill this role.

n/a

My pre-submission inquiry is surrounding the test suite. I am _very_ new to R package development and am unsure how to use testthat appropriately in my project. Examples online are very simple, and because of the spatial analysis most functions are using S4 objects which seems to complicate things further. I have attempted to find answers through my university and haven't been able to find any support. Would anyone be able to advise me on this or steer me to someone who could? I believe it is the only element that needs addressing before I can submit my package.
