### 2024-08-23 #652 @adamhsparks
Hi @heyairf, this isn't really the proper place to offer support for this request. We're happy to help but I'd suggest you try the rOpenSci forums, https://discuss.ropensci.org or if you prefer a Slack interface you could try https://dslc.io.

Since this is a pre-submission inquiry, from the looks of it, the package would be in-scope.
