### 2024-08-26 #652 @adamhsparks
If you haven't, you should refer to the Testing packages chapter in the "R-pkgs" book, https://r-pkgs.org/testing-basics.html.

A further suggestion from the editors was to check other spatial packages for their test-suites and see how they implemented them. https://github.com/r-spatialecology/landscapemetrics/tree/main/tests/testthat is one that was suggested.

I'll close this issue now. Please feel free to open a new issue for software review when you've addressed this issue with the package.
