### 2024-08-28 #652 @adamhsparks
Hi @heyairf, @maurolepore had some further suggestions that I was remiss in reporting back here. So here they are.

[21.1 Testing functions (except 21.1.4)](https://mastering-shiny.org/scaling-testing.html#testing-functions)
[21.2 Workflow](https://mastering-shiny.org/scaling-testing.html#workflow-1)
[21.6 Philosophy](https://mastering-shiny.org/scaling-testing.html#philosophy)

and Mauro also has a YouTube playlist on testing in R, https://www.youtube.com/playlist?list=PLvgdJdJDL-APxwVpLE7vjeg-Pgdd4e64W&themeRefresh=1
