### 2024-08-29 #652 @heyairf
Thank you for the additional resources! I really appreciate it. I have just posted my most pressing/specific question in the discussion forum [here](https://discuss.ropensci.org/t/s4-data-for-testthat-and-examples/4007).
